+++ Packages/ai.fxn.fxn3d/Runtime/Unity/UnityClient.cs
@@ -5,7 +5,7 @@
 
 #nullable enable
 
-namespace Function.Graph {
+namespace Function.API {
 
     using System;
     using System.Collections.Generic;
@@ -14,14 +14,15 @@ namespace Function.Graph {
     using System.Threading.Tasks;
     using UnityEngine.Networking;
     using Newtonsoft.Json;
+    using Graph;
+    using Internal;
     using Services;
 
     /// <summary>
-    /// Function graph API client for .NET.
+    /// Function API client for Unity Engine.
     /// This uses Unity APIs for performing web requests.
-    /// As such it should be used on WebGL.
     /// </summary>
-    public sealed class UnityGraphClient : IGraphClient {
+    public sealed class UnityClient : IFunctionClient {
 
         #region --Client API--
         /// <summary>
@@ -30,11 +31,11 @@ namespace Function.Graph {
         public string? Id { get; private set; }
 
         /// <summary>
-        /// Create the Unity web request client.
+        /// Create the Unity Function API client.
         /// </summary>
-        /// <param name="url">Function graph API URL.</param>
+        /// <param name="url">Function API URL.</param>
         /// <param name="accessKey">Function access key.</param>
-        public UnityGraphClient (string url, string? accessKey, string? id = null) {
+        public UnityClient (string url, string? accessKey, string? id = null) {
             this.url = url;
             this.accessKey = accessKey;
             this.Id = id;
@@ -48,23 +49,21 @@ namespace Function.Graph {
         /// <param name="input">Query inputs.</param>
         public async Task<T?> Query<T> (string query, string key, Dictionary<string, object?>? variables = default) {
             // Serialize payload
-            var payload = new GraphRequest {
-                query = query,
-                variables = variables
-            };
-            var payloadStr = JsonConvert.SerializeObject(payload);
+            var payload = new GraphRequest { query = query, variables = variables };
+            var serializationSettings = new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore };
+            var payloadStr = JsonConvert.SerializeObject(payload, serializationSettings);
             // Create client
-            using var client = new UnityWebRequest(url, UnityWebRequest.kHttpVerbPOST) {
+            using var client = new UnityWebRequest($"{this.url}/graph", UnityWebRequest.kHttpVerbPOST) {
                 uploadHandler = new UploadHandlerRaw(Encoding.UTF8.GetBytes(payloadStr)),
                 downloadHandler = new DownloadHandlerBuffer(),
                 disposeDownloadHandlerOnDispose = true,
                 disposeUploadHandlerOnDispose = true,
             };
             client.SetRequestHeader(@"Content-Type",  @"application/json");
-            // Add auth token
+            client.SetRequestHeader(@"fxn-client", Id);
             if (!string.IsNullOrEmpty(accessKey))
                 client.SetRequestHeader("Authorization", $"Bearer {accessKey}");
-            // Post
+            // Request
             client.SendWebRequest();
             while (!client.isDone)
                 await Task.Yield();
@@ -76,6 +75,42 @@ namespace Function.Graph {
                 throw new InvalidOperationException(response.errors[0].message);
             // Return
             return response.data.TryGetValue(key, out var value) ? value : default;                
+        }
+
+        /// <summary>
+        /// Perform a streaming request to a Function REST endpoint.
+        /// </summary>
+        /// <typeparam name="T">Deserialized response type.</typeparam>
+        /// <param name="path">Endpoint path.</param>
+        /// <param name="payload">POST request body.</param>
+        /// <returns>Stream of deserialized responses.</returns>
+        public async IAsyncEnumerable<T> Stream<T> (string path, Dictionary<string, object> payload) {
+            // Serialize payload
+            var serializationSettings = new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore };
+            var payloadStr = JsonConvert.SerializeObject(payload, serializationSettings);
+            // Create client
+            var downloadHandler = new DownloadHandlerAsyncIterable();
+            using var client = new UnityWebRequest($"{this.url}{path}", UnityWebRequest.kHttpVerbPOST) {
+                uploadHandler = new UploadHandlerRaw(Encoding.UTF8.GetBytes(payloadStr)),
+                downloadHandler = downloadHandler,
+                disposeDownloadHandlerOnDispose = true,
+                disposeUploadHandlerOnDispose = true,
+            };
+            client.SetRequestHeader(@"Content-Type",  @"application/json");
+            client.SetRequestHeader(@"fxn-client", Id);
+            if (!string.IsNullOrEmpty(accessKey))
+                client.SetRequestHeader(@"Authorization",  $"Bearer {accessKey}");
+            // Request
+            client.SendWebRequest();
+            // Stream
+            await foreach (var responseStr in downloadHandler.Stream()) {
+                if (client.responseCode >= 400) {
+                    var errorPayload = JsonConvert.DeserializeObject<Dictionary<string, string>>(responseStr);
+                    var error = errorPayload.TryGetValue(@"error", out var msg) ? msg : "An unknown error occurred";
+                    throw new InvalidOperationException(error);
+                }
+                yield return JsonConvert.DeserializeObject<T>(responseStr);
+            }
         }
 
         /// <summary>