+++ Changelog.md
@@ -1,3 +1,9 @@
+## 0.0.5
++ Added `Function.Predictions.Stream` method for making streaming predictions.
++ Refactored `IGraphClient` interface to `IFunctionClient`.
++ Refactored `UnityGraphClient` class to `UnityClient`.
++ Refactored `Function.Graph` namespace to `Function.API`.
+
 ## 0.0.4
 + Minor updates.
 
@@ -9,8 +15,8 @@
 + Added `FunctionUnity.ToAudioClip` extension method for converting a Function `Value` to an `AudioClip`.
 + Fixed Function access key not being available in builds.
 + Refactored `Feature` class to `Value` for improved clarity.
-+ Refactored `PredictionService.ToFeature` method to `PredictionService.ToValue`.
-+ Refactored `PredictionService.ToValue` method to `PredictionService.ToObject`.
++ Refactored `Function.Predictions.ToFeature` method to `Function.Predictions.ToValue`.
++ Refactored `Function.Predictions.ToValue` method to `Function.Predictions.ToObject`.
 + Refactored `FunctionUnity.ToFeature` extension method to `FunctionUnity.ToValue`.
 + Refactored `UploadType.Feature` enumeration member to `UploadType.Value`.